+++ src/test/java/ru/vyarus/spock/jupiter/support/ContextAccess.java
@@ -16,11 +16,17 @@ public class ContextAccess implements BeforeAllCallback, BeforeEachCallback {
         ExtensionContext root = context.getRoot();
         ActionHolder.add("root.id: "+ root.getUniqueId());
         ActionHolder.add("root.display name: " + root.getDisplayName());
+        ActionHolder.add("root.element: " + root.getElement().isPresent());
         ActionHolder.add("root.parent: " + root.getParent().isPresent());
         ActionHolder.add("root.root: " + root.getRoot().getDisplayName());
         ActionHolder.add("root.lifecycle: " + root.getTestInstanceLifecycle().isPresent());
         ActionHolder.add("root.exec mode: " + root.getExecutionMode());
         ActionHolder.add("root.exception: " + root.getExecutionException().isPresent());
+        ActionHolder.add("root.test class: " + root.getTestClass().isPresent());
+        ActionHolder.add("root.test method: " + root.getTestMethod().isPresent());
+        ActionHolder.add("root.tags: " + root.getTags());
+        ActionHolder.add("root.test instance: " + root.getTestInstance().isPresent());
+        ActionHolder.add("root.test instances: " + root.getTestInstances().isPresent());
 
         // class context
         ActionHolder.add("class.id: " + context.getUniqueId());